requirements.txt
pygame-ce==2.5.2 # Comunity Edition, works better on Apple Silicon
toml==0.10.2